Revert "Use default SPI if not specified in modem config. Thanks to @v0l (Closes: #74)"
After testing it, this commit fails on execution for most standard boards.
This project is compiled for a generic ESP32 board and the pin definitions are configured at running time, we should not rely on the standard framework pin mapping because it refers to another board and they might change!. Even if it works now it might make boards fail in the future. SPI pins MUST be declared explicitly.

This reverts commit 6bc016e47c7dd4d5aa589cf9bf5502c251a9eab1.